Match types between refs and coeffs
Duplicate av1_warped_filter table with int8_t type. Changing
type from 16-bit to 8-bit matches how data is used naturally
in the code. All values are 8-bit anyway. This enables an
incremental update of the intrinsics paths, after which we
can remove the duplicate tables and shift to 8-bit for all.

Change-Id: I03fb0e016143b38a0b7197f8b21577b470578a61