rtc: Fix to skip_encoding_non_reference_slide_change
For scene changes that occur on non_reference
frames (highest temporal layer) the encoder will do
a skip encoding. The logic to do this is bypassed
when high motion flag is set (and variance_partitioning
is skipped), which is not correct, as this can lead to
some superblocks not being skip encoded.

This fixes the artifact observed in comment#18 of the
issue below.

Bug: b:383306740
Change-Id: I35ca017edf352efe4e66a5cfbd2879ab73d8a0f6