Optimize Neon implementation of aom_lpf_6
The Neon implementations of aom_lpf_vertical_6 and aom_lpf_horizontal_6
compute both filter4() and filter6() before selecting for each element
which filter is actually needed. In practice, however, a lot of cases
only need one of the filters, so specialize for these scenarios,
computing only the filters that are needed and eliminate bitwise select.
This makes the case where all filters are needed slightly slower, but as
it is far from the most common case this is ok.

Also move the actual filter computation to separate functions to avoid
code duplication.

This is a port from d29673138e044fa7dec420f74e1451c53a9ea8ca in SVT-AV1.

Change-Id: Ibda2567c3781e51a9d43c59480284072115ceab6