Optimize Neon implementation of aom_lpf_8
The Neon implementations of aom_lpf_vertical_8 and aom_lpf_horizontal_8
compute both filter4() and filter8() before selecting for each element
which filter is actually needed. In practice, however, a lot of cases
only need one of the filters, so specialize for these scenarios,
computing only the filters that are needed and eliminate bitwise select.
This makes the case where all filters are needed slightly slower, but as
it is far from the most common case this is ok.

Also move the actual filter computation to separate functions to avoid
code duplication.

This is a port from d8ebdeec1edc1df12b2b19a6ffe477aea4bc158e in SVT-AV1.

Change-Id: I2e7157f26515f9ec51e82ad32434117adc4be928