Optimize Neon implementation of aom_lpf_14
The Neon implementations of aom_lpf_vertical_14 and
aom_lpf_horizontal_14 compute every filter (filter4(), filter8() and
filter14()) before selecting for each element which filter is actually
needed. In practice, however, a lot of cases only need one of the
filters, so specialize for these scenarios, computing only the filters
that are needed and eliminate bitwise select.  This makes the case where
all filters are needed slightly slower, but as it is far from the most
common case this is ok.

Also move the actual filter computation to separate functions to avoid
code duplication.

This is a port from 4d6f560d1f4a35a47e5dfc9b7c2dc4f5483d49dc in SVT-AV1.

Change-Id: Id608edad6fee69fc263eb6ee9e0956ca6311d77f