rtc-screen: Speed feature for fast encoding on scene change
Add speed feature to prune the newmv and palette mode testing,
to reduce the encoding time spikes, as reported in the issue below.
Spikes can be reduced by close to ~2x in some clips.

This setting has quality loss so it's added to new speed 12, and
only enabled for resolutions above 720p.
Only affects real-time mode for screen content.

Bug: b:391849810
Change-Id: Iacab126149690642faead4aa2c72ea3621c411a6