rtc: Remove num_col/row_blscroll_last_tl0
These parameters are susepct to causing the
tsan race condition in the issue below.

Removing them for now to see if the issue clears.
This may cause some slowdown in some cases of
webrtc usage, but we may add them back in when
the issue is reproduced with libaom test.

Bug: b:394563983
Change-Id: I9d1abc58ecc57921ddbbd79fb1da1dc114f974ac