Add sf adaptive_luma_filter_skip for low complexity decode
This patch introduces the speed feature 'adaptive_luma_filter_skip',
which sets 'lf->filter_level[0/1]' to zero based on the minimum filter
level of reference frames and the pyramid level of the current frame.

The logic is applied only when the current frame's luma and chroma
filter levels are below a threshold determined by its pyramid level.
Also, this logic is not applied for OVERLAY_UPDATE and
INTNL_OVERLAY_UPDATE frame types.

Change-Id: I742684b58f6835d95672c16e7b690cfb07b43c0e