lc-dec: Add sf skip_loop_filter_using_filt_error
This patch adds a speed feature 'skip_loop_filter_using_filt_error'
to reset luma filter levels based on the percentage improvement in
SSE between the unfiltered and filtered versions of the current
frame. This logic applies only when both luma and chroma improvement
in SSE is below a threshold. This sf is not enabled when frame type
is OVERLAY_UPDATE or INTNL_OVERLAY_UPDATE and current frame
pyramid_level > 1.

Performance results of encoder and decoder for the streams generated
with low complexity decode enabled.

STATS_CHANGED for low complexity decode

Change-Id: Ifae8dafa69721563ad810859ed66c3b17e5be900