rtc: Fix use of prev_frame for spatial layers
The usage of cm->prev_frame should only be for
single spatial layers, and also not used for
external_ratectrl.

Also fix the update of prev_number_spatial_layers
for dropped frames, and force the usage of
cyclic_refresh_reset_resize() only for TL0, as
cyclic is only users for TL0 frames.

Bug: b:400885218
Change-Id: Id739cbae29092736e093a56ffc0defc3760e9535