fwd_txfm2d: Highway for highbd + AVX512 lowbd
I'm not entirely satisfied with this, but it's a good enough start.

For all highbd implementations, Highway is faster across SSE4, AVX2 and
AVX512. It is unambiguously a win from a myriad of sources, but mostly
from commonizing optimizations.

For lowbd AVX512, Highway is always faster by a substantial margin
(>20%) except for TX_16X8, where on Skylake it is up to 13% slower.
There is enough of a win that I think we should simply turn it on for
all instead of special-casing TX_16X8.

For lowbd AVX2/SSE4 (not enabled here), Highway is often faster, but not
always, and so it's not enabled here. This will hopefully improve in
follow up changes.

Change-Id: I5139115cb5ef0b644b642ad3be61b55269b4b279