fwd_txfm2d: AVX512 highbd + lowbd
For lowbd AVX512, Highway is always faster by a substantial margin
(>20%) except for TX_16X8, where on Skylake it is up to 13% slower.
There is enough of a win that I think we should simply turn it on for
all instead of special-casing TX_16X8.

For lowbd AVX2/SSE4 (not enabled here), Highway is often faster, but not
always, and so it's not enabled here. This will hopefully improve in
follow up changes.

Change-Id: If7f79799041c165bd14df2e9daf1e06cf454b8f2