Fix to update seq_params for number of layers change
If the number of layers change the sequence params
(i.e., operating_points_cnt_minus_1 and operating_point_idc[])
need to be updated. The change in layers is done via the
set_svc_params control, so set seq_params_locked = 0 there
if the number of spatial or temporal layers have changed.

This fixes the assert in the bug below.
Added unittest that changes the number of temporal layers
from 1->3->1, which also triggers the assert.

Bug: b:407813259
Change-Id: I99416f0cab1cf095a62d533ddc21a7e71dcb831a