Fix to update seq_params for number of layers change
If the number of layers change the sequence params
(operating_points_cnt_minus_1 and operating_point_idc[])
need to be updated. This is done inside the set_svc_params
control, which allows for dynamic change of layers.

This fixes the assert in the bug below.

Added unittest that changes the number of temporal layers
from 1->3->1, which also triggers the assert. Also added
the check on num_operating_points in ratectrl_rtc_test.cc,
which now also triggers a failure without this fix.

Bug: b:407813259
Change-Id: I99416f0cab1cf095a62d533ddc21a7e71dcb831a