rtc: Fix to external resizing for svc
Add tests for external dynamic resizing for
spatial layers, for high resolutions (>720p).

This exposed an issue with the dynamic/default
mode for sb_size (--sb-size=0). When dynamic mode
is used the sb_size is based on resolution
(128x128 for > 720p, 64x64 otherwise), which causes
an issue when external resizing is done on the very
first frame. It can get reset from 128 to 64, which
triggers re-alloc and hence codec error when input
frame resolution goes back up dynamically.

The fix for now is for the user to set the sb-size
to either 64 or 128. For the uniittests we set:
encoder->Control(AV1E_SET_SUPERBLOCK_SIZE,
                 AOM_SUPERBLOCK_SIZE_64X64)

Also remove unneeded code (top_sl_width/height_),
and avoid unneeded check for spatial enhancement
layers that will be dropped (target_bandwidth = 0).

These issues may be the cause of the test failure in:
https://webrtc-review.googlesource.com/c/src/+/379680

Change-Id: I90491dbb1a89d8b55c5d3d7fa075606163c16e07