For spatial layers use allocated size for sb_size
This a follow-up the CL below, to remove the condition
--sb-size=64/128 for the svc external resize tests:
https://aomedia-review.googlesource.com/c/aom/+/199141

The existing condition in av1_select_sb_size() for
spatial layers (or resize) was intending to use
the allocated size (via oxcf->frm_dim_cfg.width/height),
but if the user passes in downscaled frames
(i.e., externally resize) then oxcf->frm_dim_cfg.width/height
is not the allocated size. Fix is to pass in the allocated
size directly to the select_sb_size function.

This allows for external scaling to work for spatial layers
with --sb-size=0 (default mode), and fixes the test failure in:
https://webrtc-review.googlesource.com/c/src/+/379680.

Bug: b:400885218
Change-Id: I80d32732df925cfd8a4be3a2fa3365be01f158fb