rtc: add support for ROI
Implement set_roi_map() and apply_roi_map()
to pass the user-input roi map onto encoder segments,
for the features of applying:
  -delta QP
  -delta loopfilter level
  -skip encoding
  -reference_frame selection

For now: if the active_map feature is enabled then
the roi feature is disabled. And if roi feature
is enabled then cyclic_refresh aq_mode is disabled.

Unittest and and example map input for
examples/svc_encoder_rtc.cc will be added in
subsequent CL.

Change-Id: Id4e60b13ccd509d739aaa30dd995b19de28066e4