rtc: Fix for lossless segment roi
Clamp the delta_qp so that qindex is above
zero on segment, to avoid lossless segment
case, which currently has test failures for
nonrd mode. This can triggered by setting
min_quantizer to 0 and lower max_quantizer in
the roi tests.

Allow allow for ROI delta_qp on key frames.

Change-Id: I478f00153dee464a12d4b7a9271002452477543d