rtc: For testing LTR
DO NOT COMMIT

Based on av1_discuss issue:
https://groups.google.com/a/aomedia.org/g/av1-discuss/c/sqFad980SsA

This includes a fix in the setup for the refresh flag setting
(see line 735 in svc_encoder_rtc.cc), and fixes in the nonrd_pickmode
to test non-last reference. The latter fixes to nonrd_picmode
will be made in a separate patch.

Change-Id: I5e5369e9f238d1f81c2f0038f3e98ee0ef0b00de