rtc: Fix to non_last reference in nonrd_pickmode
For nonrd_pickmode (speed >= 7):
-avoid skip testing non_last reference if LAST
is not used as a reference
-only do the fast int_pro motion estimation for
non_last reference if LAST is used as a reference.

This improves the quality for LTR case, as in this issue
from av1_discuss:
https://groups.google.com/a/aomedia.org/g/av1-discuss/c/sqFad980SsA
Follow-up will be done to further improve quality.

No change in rtc test stats as LAST is always used as a
reference there.

Change-Id: Ic20d202b512cd903e9b47a389a905f75f933e8ec