Allow file paths with 'weird' characters.
Currently, if e.g. `AOM_CONFIG_DIR` is set to a path with special characters (such as `+`), the cmake build fails. This is because `MATCHES` interprets the second string as a regex, which is invalid.

Since we only care about substring existance in these checks, rewrite to using `FIND`. Unfortunately, I could not find a way to do this in a single line, so changed to two lines. Since I am very much not a cmake expert, I'd appreciate advice for improving the syntax.

This is specifically a problem for us when using Bazel to build aom (with the help of rules_foreign_cc, so essentially wrapping cmake in Bazel). Bazel uses `+` in resolved paths for external dependencies and our build is currently failing with

```
CMake Error at CMakeLists.txt:1037 (if):
  if given arguments:

    "NOT" "/usr/local/google/home/alexanderfaxa/.cache/bazel/_bazel_alexanderfaxa/aa7e559759091793a65d16afaac928ae/sandbox/linux-sandbox/281/execroot/_main/external/module++non_module_deps_ext+libaom/av1/common/x86/av1_inv_txfm_avx2.c" "MATCHES" "/usr/local/google/home/alexanderfaxa/.cache/bazel/_bazel_alexanderfaxa/aa7e559759091793a65d16afaac928ae/sandbox/linux-sandbox/281/execroot/_main/bazel-out/haswell-opt/bin/external/module++non_module_deps_ext+libaom/libaom.build_tmpdir"

  Regular expression
  "/usr/local/google/home/alexanderfaxa/.cache/bazel/_bazel_alexanderfaxa/aa7e559759091793a65d16afaac928ae/sandbox/linux-sandbox/281/execroot/_main/bazel-out/haswell-opt/bimodulen/external/module++non_module_deps_ext+libaom/libaom.build_tmpdir"
  cannot compile

RegularExpression::compile(): Nested *?+.
RegularExpression::compile(): Error in compile.
```

Change-Id: I1b8a1807d4c438312a16805cfe2b6d4d19dd1572