rtc: Restrict usage of src_sad_blk_64x64 to single spatial layer
For blk_sad usage: disallow for spatial layers.

This is a possible fix for crash in the issue below.

The issue seems to be incorrect setting of svc->mi_cols/rows_full_resoln (used for the allocation of src_sad_blk_64x64) for the case of switching up layers (2SL to 3SL) with passing in lower resolution for the top layer after the switch.

The performance effect of src_sad_blk_64x64 for spatial layered encoding is expected to be small/minimal, so ok to disable. Will follow-up with a unittest and may re-enable the feature with the correct fix.

Bug b:433046392, b:432035817

Change-Id: Id0b2b36771fffef75bc89f51cf62709d5b6ad606