Improve comment for AOM_BUFFER_SIZE_FOR_BLOCK_HASH
The original definition "Number of 2x2 pixel blocks per superblock" may
be misinterpreted to mean that this buffer is only used to hold the hash
values of 2x2 subblocks in a superblock. The new definition clarifies
that the subblock size and the block size are not fixed. It also avoids
mentioning superblock and just says largest intraBC block size is
128x128.

Bug: aomedia:401234702

Change-Id: I445f80829110fb188e20456782606c47bebdf992