Disable access for 'Blocked Users' group
This may fix the issue with the global setting as the block (for Blocked
Users) and the allow (for Registered Users) are in the same section:

https://gerrit-review.googlesource.com/Documentation/access-control.html

Note: An 'ALLOW' rule overrides a 'BLOCK' rule only when both of them
are inside the same access section of the same project. An 'ALLOW' rule
in a different access section of the same project or in any access
section in an inheriting project cannot override a 'BLOCK' rule.

Change-Id: I0d964bb36f82b9130a0574c3042ff7f50fa95c5e