feat: Add AOM_EFLAG_FREEZE_INTERNAL_STATE
Add a flag to encode a frame without updating encoder state.
This is useful for speculative encoding.

The flag ensures the current frame is encoded normally, PSNR is
generated if requested, but internal state like reference buffers,
CDF tables, and rate control is not modified.

Adds tests for single layer and SVC encoding.

Change-Id: I34e9e3d0beb355cff2f9b3419820773e5d8c550d