Optimize av1_dr_prediction_z1/3 Neon implementation
Replace multiplication of 32 to left shift of 5 for lower instruction
latency and to utilize the other 2 SIMD pipelines.

Also remove the never used upsample_above=1 case in
dr_prediction_z1_WxH_internal_neon_large, as according to
av1_use_intra_edge_upsample the upsample_above is 0 for when bw+bh>16.

This optimization is a port from SVT-AV1.

Change-Id: Ifcbc2e5f8349c1ecf3d3cc09dda73032ea5d167c