rtc-.screen: Rework the logic for setting the high motion flag
Move the reset of rc->num_col/row_roll_blscroll_last_tl0 to after
the frame dropping (so it corresponds to the last encoded frame)
when used in the scene_detection. And add condition on the last
frame where the source_sad was low. This improves the setting of
rc->high_motion_screen_content_rtc,

This fixes the encode_time regression noted in the issue below,
and still has the reduced quality dips that was fixed in:
https://aomedia-review.googlesource.com/c/aom/+/203801

Bug: b:444116924
Change-Id: Ie63e54228b1616341a802e525b78601dc7dc92da