v256_intrinsics_c.h: fix AVX2/C mismatch w/gcc+ubsan
This also clears the `-Wmaybe-uninitialized` warnings with gcc and
`-fsanitize=undefined`.

This is a backport of the libavm change:

7f95277d98 Fix AVX/C mismatch bug for v256_wideshuffle_8

Fix AVX/C mismatch bug for v256_wideshuffle_8

The bug was not in SIMD. In the reference C implementation array
indexing had a ternary operator as a clever shortcut.
This operation confused GCC compiler which produced different results
compared to SIMD that was correct, clang also matches SIMD correctly.
Fix explicitly defines the index as int outside the array.

This caused the reported bug in https://gitlab.com/AOMediaCodec/avm/-/issues/95.

Bug: 42302489
Change-Id: I0c3e09275db4d004f3669bd739cfbeb956a3dbdb