Realtime mode with lookahead, 1 pass vbr
Make nonzero lookahead (lag_in_frames > 0) work with
1 pass realtime mode. This new encoding mode is for
single spatial layer and VBR mode.

The gop/pyramid strcture is set in the function
define_gf_group_pass0(), and the gop length and depth
is set by the user via the parameters:
--min/max-gf-interval, --gf-min/max-pyr-height

The gop structure will be fixed (based on user input) for
the whole sequence for now, adaptation to the source content
will be added later.

Nonrd-pickmode (speed >= 7) is updated to support the
additional references for prediction in this mode,

Test for this encoding mode added in datarate_test,cc.

The case of spatial layers was done here (not merged in):
https://aomedia-review.googlesource.com/c/aom/+/193401

Change-Id: I50eeff1a87884f7237f5163b209f79fdae899979