Realtime mode with lookahead for 1 pass
Make nonzero lookahead (lag_in_frames > 0) work with
1 pass realtime mode. This new encoding mode is for
single spatial layer and VBR mode. And disallowed
for REALTIIME_CONFIG=1 (lag-in-frames is kept forced
to 0 for this build config under realtime mode).

The gop/pyramid structure is set in the function
define_gf_group_pass0(), with the gop length and depth
set by the user via the parameters:
--min/max-gf-interval, --gf-min/max-pyr-height

The gop structure will be fixed (based on user settings) for
the whole sequence for now, adaptation to the source content
will be added later.

The nonrd-pickmode encoding (speed >= 7) is expanded to
support the additional references for prediction in this
encoding mode.

Test for this encoding mode added in datarate_test,cc.

Initial borg result for speed 8, comparing nonzero lookahead
(with gf-interval=16 and pyr-height=1) vs lookahead=0:
bdrate gain of ~15% with IC speed slowdown of ~50%.

The case of spatial layers was done here (not merged in):
https://aomedia-review.googlesource.com/c/aom/+/193401

Change-Id: I50eeff1a87884f7237f5163b209f79fdae899979