Allow for realtime mode with lookahead
Make lag_in_frames > 0 work with realtime mode.
This new encoding mode is for single spatial layer
and VBR mode. The layered/pyramid strcture is set in
the function define_gf_group_pass0(), and needs to be
modified/updated.

The case of spatial layers was done here (not merged in):
https://aomedia-review.googlesource.com/c/aom/+/193401

Change-Id: I50eeff1a87884f7237f5163b209f79fdae899979