rtc: Extend rtc test coverage down to speed 5
Extend rtc test coverage down to speed 5, with
additional tests. This is also an attempt to trigger
the crash in the issue below, which is rtc encoding,
using lower speed settings (5 or 6) for rd pickmode.

Also remove the ResetModel() in
RunBasicRateTargetingTestReversed(), as this
will override some intended settings for the
tests.

Bug: b:457951958
Change-Id: I6131c61589f5012063de860a44f430391209828d