Speculative fix for issue b:457951958
Crash happens for speed 6 realtime encoding,
and points to mi->bsize being invalid. For now try
a fix of using bsize instead of mi->bsize for
the mi_size_width/height[].

The check for (mi != NULL) is done in parent functions
where the pc_tree is allocated, so issue is likely
that mi->bsize (ctx->mic.bsize) is not set.

Bug: b:457951958
Change-Id: I80f3a3c1ea90d67e08017c3a4eaa13bb18ff811b