Speculative fix for issue b:457951958
Crash happens for speed 6 realtime encoding, and
points to mi->bsize being invalid in av1_update_state().
For now try a fix of using bsize instead of mi->bsize
for the mi_size_width/height[].

Add check and return codec_error if mi == NULL in
av1_update_state(). The check for (mi != NULL) is already done
in the parent functions where the pc_tree is allocated, but
add the additional check to be sure.

Bug: b:457951958
Change-Id: I80f3a3c1ea90d67e08017c3a4eaa13bb18ff811b