rtc: Speculative fix for issue; 457951958
For realtime mode in rd_pick_intra: set the mode defaults
before entering pickmode. This is to guard against the case
where an early return (due to rd_cost->rate == INT_MAX)
happens before any mode is set, in which case no mode parameters
are set. For realtime mode the rd_pick_intra is used for
speeds 5 and 6 on all frames, and for small blocks on keyframes for speeds >= 7 (nonrd mode).

This is done as a possible fix to a crash issue attached below.
But we have not been able to reproduce this crash or reproduce
this type of early exit. So this is speculative fix.

Bit-exact/no stats changed on all rtc sets.

Bug: b:457951958

Change-Id: I5c7e116b94d1734690772505d524e33639ddd6b9